ai/llm-agents/langgraf-openai-tavily/pyproject.toml
[project]
name = "43-langgraf"
version = "0.1.0"
description = "LangGraf demo"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ddgs>=9.11.2",
    "duckduckgo-search>=8.1.1",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.10",
    "langchain-tavily>=0.2.17",
    "langgraph>=1.0.10",
    "python-dotenv>=1.2.2",
    "tavily-python>=0.7.22",
]
